Refactor to train multiple configs

+++ chemdner_annotate.sh
@@ -1,7 +1,13 @@
 #! /bin/bash
 
+model_version="$1"
+if [ -z $model_version ]; then
+    echo "Please specify the version of the model which you want to annotate (e.g. v0, v1, etc) "
+    exit
+fi
+
 rm -rf "./chemdner_corpus/annotate/in/" && mkdir -p "./chemdner_corpus/annotate/in/"
-echo "Preparing annotation input..." && sleep 1
+echo "[$model_version] Preparing annotation input..." && sleep 1
 while IFS=$'\t' read -r field1 field2 field3; do
     echo "$field1"
     echo -e "$field2" > "./chemdner_corpus/annotate/in/"$field1"T.txt"
@@ -9,15 +15,15 @@ while IFS=$'\t' read -r field1 field2 field3; do
 done < ./chemdner_corpus/development.abstracts.txt
 
 rm -rf "./chemdner_corpus/annotate/out/" && mkdir -p "./chemdner_corpus/annotate/out/"
-echo "Annotating..." && sleep 1
+echo "[$model_version] Annotating..." && sleep 1
 ./neji.sh -i chemdner_corpus/annotate/in/ -if RAW \
           -o chemdner_corpus/annotate/out/ -of A1 \
           -m enumerated_suffix_prefix/ \
           -v \
           -noids
 
-rm "./chemdner_corpus/predictions.txt"
-echo "Building single prediction file..." && sleep 1
+rm "./chemdner_corpus/predictions_$model_version.txt"
+echo "[$model_version] Building single prediction file..." && sleep 1
 for filename in $(find ./chemdner_corpus/annotate/out/ -type f -not -empty); do
     echo $filename
     id=$(basename $filename | cut -d'.' -f1 | head -c -2)
@@ -26,7 +32,7 @@ for filename in $(find ./chemdner_corpus/annotate/out/ -type f -not -empty); do
     while IFS=$'\t' read -r field1 field2 field3; do
         startidx=$(echo $field2 | cut -d " " -f2)
         endidx=$(echo $field2 | cut -d " " -f3)
-        echo -e "$id\t$sentence_type:$startidx:$endidx" >> ./chemdner_corpus/predictions.txt
+        echo -e "$id\t$sentence_type:$startidx:$endidx" >> "./chemdner_corpus/predictions_$model_version.txt"
         read
     done < "$filename"
 done
